Package Aquilon
Add a setup.py and a few scripts to build RPMs that can be shipped out
of Morgan Stanley.  The version is derived from git-describe, to
simplify the release process.

The only daemon packaged the Red Hat variant.

All dependencies to modules internal to MS are removed during the %prep
phase.  All the declared dependencies have either RPMs in SL or they are
easy to build (and are being documented in quattor.org).

Configuration files with default values are installed into
/usr/share/aquilon/etc.  Templates for mako go into
/usr/share/aquilon/mako.

During the %build phase we process the bash autocompletion script.  We
also ship a %post script, that will regenerate the twisted caches, to
avoid annoying exceptions.

Also, some needed Python modules are located in the build/ directory,
and distutils won't package them.  The setup.py cheats by creating a
symlink and referring to it during the build process.