Build and package all the documentation
Provides variants of the style-html and style-man XSLs that include the
XSLs supplied by Red Hat and derivatives.

The Makefile is adapted to take as a parameter the style of the XSL to
be used.  It defaults to the Morgan Stanley one.  The same file can now
build in Morgan Stanley and in RPM environments.

The build-autocomplete now also builds all the documentation.

The setup.py packages the docs as well, and the setup.cfg lists the new
build dependencies.